requirements.txt
beautifulsoup4
gspread==6.2.1
gspread_dataframe==4.0.0
lxml==4.9.3
oauth2client==4.1.3
pandas==2.2.3
pyautogui==0.9.54
PyDrive2==1.21.3
python-dotenv==1.1.0
python-telegram-bot==22.1
python_dateutil==2.8.2
selenium
seleniumbase
pyotp
